Move test suite to new file
